refactor(train): represent route basePrice with Money and long
Unify money handling by introducing a Money value object and persist amounts as
the lowest denomination (long) to simplify conversions and match DB constraints.

- Add Money.vnd factory and toLong to convert amounts to the lowest denomination
- Change Route.basePrice from BigDecimal to Money to encapsulate currency
  semantics
- Update DTOs, HTTP requests/responses, and mappers to use long for persistence
  and API payloads to simplify serialization
- Update entity and repository mapper to store base_price as long (BIGINT) to
  align with DB schema and constraints
- Update use cases to convert Money.toLong() when producing DTOs for output
- Update RouteStatus enum to include BOARDING, DEPARTED, and ARRIVED to match DB
  lifecycle constraint
- Update tests to use Money.vnd and assert long basePrice values to reflect the
  new representation

BREAKING CHANGE: base_price column type changed from DECIMAL to BIGINT and API
DTOs now use long for basePrice

+++ backend/src/main/java/io/github/phunguy65/ttbs/backend/train/application/dto/RouteDto.java
@@ -1,7 +1,6 @@
 package io.github.phunguy65.ttbs.backend.train.application.dto;
 
 import io.github.phunguy65.ttbs.backend.train.domain.model.RouteStatus;
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -12,6 +11,6 @@ public record RouteDto(
         UUID destinationStationId,
         Instant departureTime,
         Instant arrivalTime,
-        BigDecimal basePrice,
+        long basePrice,
         RouteStatus status,
         Instant createdAt) {}
