+++ backend/src/main/java/io/github/phunguy65/ttbs/backend/train/application/usecase/CreateRouteUseCase.java
@@ -57,7 +57,7 @@ public class CreateRouteUseCase {
                 route.getDestinationStationId().value(),
                 route.getDepartureTime(),
                 route.getArrivalTime(),
-                route.getBasePrice(),
+                route.getBasePrice().toLong(),
                 route.getStatus(),
                 route.getCreatedAt());
     }
