+++ backend/src/main/java/io/github/phunguy65/ttbs/backend/train/application/usecase/GetRouteByIdUseCase.java
@@ -34,7 +34,7 @@ public class GetRouteByIdUseCase {
                 route.getDestinationStationId().value(),
                 route.getDepartureTime(),
                 route.getArrivalTime(),
-                route.getBasePrice(),
+                route.getBasePrice().toLong(),
                 route.getStatus(),
                 route.getCreatedAt());
     }
