+++ backend/src/main/java/io/github/phunguy65/ttbs/backend/train/application/usecase/GetRoutesUseCase.java
@@ -38,7 +38,7 @@ public class GetRoutesUseCase {
                 route.getDestinationStationId().value(),
                 route.getDepartureTime(),
                 route.getArrivalTime(),
-                route.getBasePrice(),
+                route.getBasePrice().toLong(),
                 route.getStatus(),
                 route.getCreatedAt());
     }
