+++ backend/src/main/java/io/github/phunguy65/ttbs/backend/train/infrastructure/web/RouteHttpResponse.java
@@ -1,7 +1,6 @@
 package io.github.phunguy65.ttbs.backend.train.infrastructure.web;
 
 import io.github.phunguy65.ttbs.backend.train.domain.model.RouteStatus;
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -12,6 +11,6 @@ record RouteHttpResponse(
         UUID destinationStationId,
         Instant departureTime,
         Instant arrivalTime,
-        BigDecimal basePrice,
+        long basePrice,
         RouteStatus status,
         Instant createdAt) {}
