+++ backend/src/main/java/io/github/phunguy65/ttbs/backend/train/infrastructure/web/UpdateRouteHttpRequest.java
@@ -2,14 +2,13 @@ package io.github.phunguy65.ttbs.backend.train.infrastructure.web;
 
 import io.github.phunguy65.ttbs.backend.train.domain.model.RouteStatus;
 import jakarta.validation.constraints.PositiveOrZero;
-import java.math.BigDecimal;
 import java.time.Instant;
 import org.openapitools.jackson.nullable.JsonNullable;
 
 record UpdateRouteHttpRequest(
         JsonNullable<Instant> departureTime,
         JsonNullable<Instant> arrivalTime,
-        @PositiveOrZero JsonNullable<BigDecimal> basePrice,
+        @PositiveOrZero JsonNullable<Long> basePrice,
         JsonNullable<RouteStatus> status) {
 
     UpdateRouteHttpRequest() {
